Allow ingester max-exemplars setting to be changed dynamically
This allows admins to change the limit without restarting ingesters.
Do it for all tenants, periodically, because that is a very simple
implementation and covers cases like the number of ingesters changing.

As part of this, move max exemplars setting to Overrides.

Also make it a global value across the cluster, which gets divided by
the number of ingesters to compute the limit used locally.

Signed-off-by: Bryan Boreham <bjboreham@gmail.com>