Update memcached, minio, etcd, consul and dynamodb e2e images
* Update minio, etcd, consul and dynamodb e2e images

Current minio and dynamodb images do not have an arm64 version. Because
of this docker desktop on m1 macs complains.

Bumping memcached, etcd and consul versions for good measure since it's
only patch releases.

* Downgrade minio and update healthcheck EP

Minio changed the semantics of the /ready endpoint, so we have to use
/cluster to get write readiness. The cluster write readiness
endpoint changed behaviour some time between feb and august this year
(I think unintentionally, but don't have more time to spend on this),
that's why the update is not to the latest image.